+++ app/service/auth/auth.ts
@@ -8,13 +8,32 @@ import {
 import { User } from '../../models/User.js';
 import logger from '../../utils/logger.js';
 
-export const createUser = async (
-  req: Request<{ User: User }>,
-  res: Response
-) => {
+export const createUser = async (req: Request, res: Response) => {
   try {
     const data = req?.body;
+    const { name, email, password, role } = data;
+
+    if (!name && !email && !password && !role) {
+      res
+        .status(400)
+        .json({ error: 'name,email,password,role fields are required' });
+      return;
+    } else if (!name || name.trim() === '') {
+      res.status(400).json({ error: 'Name is required' });
+      return;
+    } else if (!email || email.trim() === '') {
+      res.status(400).json({ error: 'Email is required' });
+      return;
+    } else if (!password || password.trim() === '') {
+      res.status(400).json({ error: 'Password is required' });
+      return;
+    } else if (!role || role.trim() === '') {
+      res.status(400).json({ error: 'Role is required' });
+      return;
+    }
+
     const id = await addUser(data);
+    
     res.status(201).json({ id, message: 'User created successfully' });
   } catch (err) {
     logger.error('Error in creating account: ', err);
@@ -22,44 +41,85 @@ export const createUser = async (
   }
 };
 
-export const getUserById = async (
-  req: Request<{ uid: string }>,
-  res: Response
-) => {
+export const getUserById = async (req: Request, res: Response) => {
   try {
     const uid = req.params?.uid;
+
+    if (!uid || uid.trim() === '') {
+      res.status(400).json({ error: 'User ID is required' });
+      return;
+    }
+
     const user: User = await getUserByIdApi(uid);
     res.status(200).json({ user: user });
-  } catch (err) {
-    logger.error('Error in fetching user: ', err);
+  } catch (err: unknown) {
+    logger.error('Error in fetching user: ', err as Error);
+
+    const e = err as { code?: string; httpCode?: number; message?: string };
+
+    if (e?.code === 'auth/user-not-found') {
+      res.status(404).json({
+        error:
+          'There is no user record corresponding to the provided identifier.'
+      });
+      return;
+    }
+
+    if (e?.httpCode && e?.message === 'User data not found in database.') {
+      res.status(e.httpCode).json({ error: e.message });
+      return;
+    }
+
     res.status(500).json({ error: 'Internal Server Error' });
   }
 };
 
-export const validateEmail = async (
-  req: Request<{ email: string }>,
-  res: Response
-) => {
+export const validateEmail = async (req: Request, res: Response) => {
   try {
     const email = req.body?.email;
+    if (!email || email.trim() === '') {
+      res.status(400).json({ error: 'Email ID is required' });
+      return;
+    }
     const isValid = await validateEmailApi(email);
     res.status(200).json({ isValid: isValid });
-  } catch (err) {
-    logger.error('Error in validating email: ', err);
+  } catch (err: unknown) {
+    logger.error('Error in validating email: ', err as Error);
+    const e = err as { code?: string; httpCode?: number; message?: string };
+    if (e?.code === 'auth/user-not-found') {
+      res.status(404).json({ error: 'User email is not found.' });
+      return;
+    }
+    if (e?.httpCode && e?.message === 'User Email not found.') {
+      res.status(e.httpCode).json({ error: e.message });
+      return;
+    }
     res.status(500).json({ error: 'Internal Server Error' });
   }
 };
 
-export const deleteUser = async (
-  req: Request<{ uid: string }>,
-  res: Response
-) => {
+export const deleteUser = async (req: Request, res: Response) => {
   try {
     const uid = req.params?.uid;
+    if (!uid || uid.trim() === '') {
+      res.status(400).json({ error: 'User ID is required' });
+      return;
+    }
     await deleteUserApi(uid);
     res.status(200).json({ message: 'User deleted successfully' });
   } catch (err) {
-    logger.error('Error in deleting user: ', err);
+    logger.error('Error in deleting user: ', err as Error);
+    const e = err as { code?: string; httpCode?: number; message?: string };
+    if (e?.code === 'auth/user-not-found') {
+      res
+        .status(404)
+        .json({ error: 'Unable to delete the user, User not found.' });
+      return;
+    }
+    if (e?.httpCode && e?.message === 'User data not found.') {
+      res.status(e.httpCode).json({ error: e.message });
+      return;
+    }
     res.status(500).json({ error: 'Internal Server Error' });
   }
 };
